wip: Keep bare minimum to build for dafang
This work adds the needed files to compile with the Dafang-Hacks codebase.

It works, but is done very sloppy.

Currently every access to sqlite3 is removed (missing library).
The overlay stuff is also removed (not needed for my purpose, reduce potential cause of crashes).
Update imp headers taken from the videoserver iirc.

This work is mainly pushed in case anyone else wants to try this on a Ingenic based camera using the Dafang Hacks framework (less hardware requirements).

Signed-off-by: Darrel Griët <dgriet@gmail.com>